Refactor functions out of brazeBanner.ts
